Don’t use python’s os.path.join.
When os.path.join finds an absolute path in the argument list it ignores 
the other parameters and just uses the absolute path. I think this is 
aimed at people who are constructing ambiguous paths out of CWD and 
a maybe absolute path or relative path. 

In any case, we do not want this behavior.