Extend mods to have root and data parameters.
Root specifies where the farm will be created.
Data specifies where the User Data Directory will be.

Volume changed to have a ‘udd’ path key. Used to find the blob store.

This allows for UDD to be outside of the farm itself.
This will enable sharing of UDD between farms.

Remove makePath function: This function was str -> IO Path.
The IO is because it inferred the CWD. This is a bad pattern,
as it wasn’t clear that callers were doing IO. Now the UI explicitly 
builds its paths with CWD and the Path constructor.